+++ src/shims/mod.rs
@@ -1,4 +1,4 @@
-
+mod backtrace;
 pub mod foreign_items;
 pub mod intrinsics;
 pub mod posix;
